🗑️&♻️ :: 불필요한 코드&파일 삭제 & 파일명 변경
- 사용되지 않는 레거시 코드 및 파일 삭제
- 파일명 변경 등 리팩토링 작업
