+++ Localizable.xcstrings
@@ -17,16 +17,6 @@
         }
       }
     },
-    "%@ %@" : {
-      "localizations" : {
-        "ko" : {
-          "stringUnit" : {
-            "state" : "new",
-            "value" : "%1$@ %2$@"
-          }
-        }
-      }
-    },
     "%@년" : {
 
     },
@@ -96,9 +86,6 @@
     "부정" : {
 
     },
-    "삭제" : {
-
-    },
     "슛-쏘기" : {
 
     },
@@ -127,9 +114,6 @@
 
     },
     "전체 (%lld개)" : {
-
-    },
-    "정말 삭제하시겠습니까?" : {
 
     },
     "지금 감정을 입력해봐요" : {
