--- Rebound Journal/Screen/Photo/PhotoPicker.swift
@@ -1,42 +1,0 @@
-//
-//  PhotoPicker.swift
-//  Rebound Journal
-//
-//  Created by hyunho lee on 12/8/24.
-//
-
-import SwiftUI
-import PhotosUI
-
-/// Generic photo picker
-struct PhotoPicker: UIViewControllerRepresentable {
-    
-    @Environment(\.presentationMode) var presentationMode
-    let didSelect: (_ image: UIImage?) -> Void
-
-    func makeUIViewController(context: UIViewControllerRepresentableContext<PhotoPicker>) -> UIImagePickerController {
-        let picker = UIImagePickerController()
-        picker.delegate = context.coordinator
-        picker.allowsEditing = false
-        return picker
-    }
-    
-    func updateUIViewController(_ uiViewController: UIImagePickerController, context: Context) { }
-    
-    func makeCoordinator() -> Coordinator {
-        Coordinator(self)
-    }
-    
-    class Coordinator: NSObject, UINavigationControllerDelegate, UIImagePickerControllerDelegate {
-        let parent: PhotoPicker
-
-        init(_ parent: PhotoPicker) {
-            self.parent = parent
-        }
-        
-        func imagePickerController(_ picker: UIImagePickerController, didFinishPickingMediaWithInfo info: [UIImagePickerController.InfoKey: Any]) {
-            parent.didSelect((info[.originalImage] as? UIImage))
-            parent.presentationMode.wrappedValue.dismiss()
-        }
-    }
-}
